+++ algorithms_datastructures/graphs/traversal/linked_stack.py
